// src/app/generated/api/model/list.ts
/**
 * Listok API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Item } from './item';


/**
 * A List containing multiple Items.
 */
export interface List { 
    /**
     * Name of the List.
     */
    name: string;
    /**
     * Optional comment about the List.
     */
    comment?: string;
    /**
     * Optional image URL for the List.
     */
    img?: string;
    /**
     * Unique identifier for the List.
     */
    readonly id: string;
    /**
     * The list of Items.
     */
    items: Array<Item>;
}